requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
jinja2==3.1.2
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.4.2